Move download.get_file_content to req.req_file
